+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/PopulationStatisticsJob.java
@@ -17,12 +17,9 @@ package uk.ac.ebi.eva.pipeline.jobs;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.batch.core.BatchStatus;
 import org.springframework.batch.core.Job;
-import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
-import org.springframework.batch.core.job.builder.FlowBuilder;
 import org.springframework.batch.core.job.builder.JobBuilder;
 import org.springframework.batch.core.job.flow.Flow;
 import org.springframework.batch.core.launch.support.RunIdIncrementer;
@@ -31,9 +28,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
-import uk.ac.ebi.eva.pipeline.jobs.deciders.SkipStepDecider;
-import uk.ac.ebi.eva.pipeline.jobs.steps.PopulationStatisticsGeneratorStep;
-import uk.ac.ebi.eva.pipeline.jobs.steps.PopulationStatisticsLoaderStep;
+import uk.ac.ebi.eva.pipeline.jobs.flows.PopulationStatisticsFlow;
 
 /**
  * Configuration to run a full Statistics job: variantStatsFlow: statsCreate --> statsLoad
@@ -41,25 +36,18 @@ import uk.ac.ebi.eva.pipeline.jobs.steps.PopulationStatisticsLoaderStep;
  */
 @Configuration
 @EnableBatchProcessing
-@Import({PopulationStatisticsGeneratorStep.class, PopulationStatisticsLoaderStep.class})
+@Import({PopulationStatisticsFlow.class})
 public class PopulationStatisticsJob extends CommonJobStepInitialization{
 
     private static final Logger logger = LoggerFactory.getLogger(PopulationStatisticsJob.class);
     private static final String jobName = "calculate-statistics";
-    public static final String SKIP_STATS = "statistics.skip";
-    public static final String CALCULATE_STATISTICS = "Calculate statistics";
-    public static final String LOAD_STATISTICS = "Load statistics";
-    private static final String STATS_FLOW = "statsFlow";
 
     @Autowired
     private JobBuilderFactory jobBuilderFactory;
 
     @Autowired
-    private PopulationStatisticsGeneratorStep populationStatisticsGeneratorStep;
-
-    @Autowired
-    private PopulationStatisticsLoaderStep populationStatisticsLoaderStep;
-
+    private Flow optionalStatisticsFlow;
+    
     @Bean
     public Job variantStatisticsJob() {
         JobBuilder jobBuilder = jobBuilderFactory
@@ -67,27 +55,8 @@ public class PopulationStatisticsJob extends CommonJobStepInitialization{
                 .incrementer(new RunIdIncrementer());
 
         return jobBuilder
-                .start(optionalStatisticsFlow())
+                .start(optionalStatisticsFlow)
                 .build().build();
     }
 
-    @Bean
-    public Flow optionalStatisticsFlow(){
-        SkipStepDecider statisticsSkipStepDecider = new SkipStepDecider(getPipelineOptions(), SKIP_STATS);
-
-        return new FlowBuilder<Flow>(STATS_FLOW)
-                .start(statisticsSkipStepDecider).on(SkipStepDecider.DO_STEP)
-                .to(statsCreate())
-                .next(statsLoad())
-                .from(statisticsSkipStepDecider).on(SkipStepDecider.SKIP_STEP).end(BatchStatus.COMPLETED.toString())
-                .build();
-    }
-
-    private Step statsCreate() {
-        return generateStep(CALCULATE_STATISTICS, populationStatisticsGeneratorStep);
-    }
-
-    private Step statsLoad() {
-        return generateStep(LOAD_STATISTICS, populationStatisticsLoaderStep);
-    }
 }
